requirements.txt
Pillow
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp37-cp37m-linux_x86_64.whl
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
cryptography==2.6.1
future==0.17.1
pipenv==2018.11.26
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==12.0.0b1
six==1.12.0
tornado==6.0.2
virtualenv==16.4.3
virtualenv-clone==0.5.1
wincertstore==0.2
fastai
